Revert "Extract out YamlCompiler"
This reverts commit 4a183776d3606e766ce6b9e3b552309d385bb716.
