+++ github-actions/src/main/scala/org/typelevel/sbt/gha/GenerativePlugin.scala
@@ -63,7 +63,54 @@ object GenerativePlugin extends AutoPlugin {
   }
 
   import autoImport._
-  import YamlCompiler._
+
+  private def indent(output: String, level: Int): String = {
+    val space = (0 until level * 2).map(_ => ' ').mkString
+    (space + output.replace("\n", s"\n$space")).replaceAll("""\n[ ]+\n""", "\n\n")
+  }
+
+  private def isSafeString(str: String): Boolean =
+    !(str.indexOf(':') >= 0 || // pretend colon is illegal everywhere for simplicity
+      str.indexOf('#') >= 0 || // same for comment
+      str.indexOf('!') == 0 ||
+      str.indexOf('*') == 0 ||
+      str.indexOf('-') == 0 ||
+      str.indexOf('?') == 0 ||
+      str.indexOf('{') == 0 ||
+      str.indexOf('}') == 0 ||
+      str.indexOf('[') == 0 ||
+      str.indexOf(']') == 0 ||
+      str.indexOf(',') == 0 ||
+      str.indexOf('|') == 0 ||
+      str.indexOf('>') == 0 ||
+      str.indexOf('@') == 0 ||
+      str.indexOf('`') == 0 ||
+      str.indexOf('"') == 0 ||
+      str.indexOf('\'') == 0 ||
+      str.indexOf('&') == 0)
+
+  private def wrap(str: String): String =
+    if (str.indexOf('\n') >= 0)
+      "|\n" + indent(str, 1)
+    else if (isSafeString(str))
+      str
+    else
+      s"'${str.replace("'", "''")}'"
+
+  def compileList(items: List[String], level: Int): String = {
+    val rendered = items.map(wrap)
+    if (rendered.map(_.length).sum < 40) // just arbitrarily...
+      rendered.mkString(" [", ", ", "]")
+    else
+      "\n" + indent(rendered.map("- " + _).mkString("\n"), level)
+  }
+
+  def compileListOfSimpleDicts(items: List[Map[String, String]]): String =
+    items map { dict =>
+      val rendered = dict map { case (key, value) => s"$key: $value" } mkString "\n"
+
+      "-" + indent(rendered, 1).substring(1)
+    } mkString "\n"
 
   def compilePREventType(tpe: PREventType): String = {
     import PREventType._
